fix: resolve WebSocket connection failure for Nyay Saarthi (#465)
