+++ backend/nyaysetu-backend/src/main/java/com/nyaysetu/backend/config/StompWebSocketConfig.java
@@ -1,0 +1,27 @@
+package com.nyaysetu.backend.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.messaging.simp.config.MessageBrokerRegistry;
+import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
+import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
+import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
+
+@Configuration
+@EnableWebSocketMessageBroker
+public class StompWebSocketConfig implements WebSocketMessageBrokerConfigurer {
+
+    @Override
+    public void configureMessageBroker(MessageBrokerRegistry config) {
+        config.enableSimpleBroker("/topic", "/queue");
+        config.setApplicationDestinationPrefixes("/app");
+    }
+
+    @Override
+    public void registerStompEndpoints(StompEndpointRegistry registry) {
+        registry.addEndpoint("/api/ws/stomp")
+                .setAllowedOriginPatterns("*")
+                .withSockJS();
+        registry.addEndpoint("/api/ws/stomp")
+                .setAllowedOriginPatterns("*");
+    }
+}
