+++ backend/nyaysetu-backend/src/main/java/com/nyaysetu/backend/config/WebSocketConfig.java
@@ -1,50 +1,22 @@
 package com.nyaysetu.backend.config;
 
+import com.nyaysetu.backend.handler.NotificationWebSocketHandler;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-import com.nyaysetu.backend.handler.NotificationWebSocketHandler;
-
-import lombok.RequiredArgsConstructor;
 
 @Configuration
 @EnableWebSocket
-@EnableWebSocketMessageBroker
 @RequiredArgsConstructor
-public class WebSocketConfig implements WebSocketConfigurer, WebSocketMessageBrokerConfigurer {
+public class WebSocketConfig implements WebSocketConfigurer {
 
     private final NotificationWebSocketHandler notificationHandler;
 
-    // Raw WebSocket handler for notifications
     @Override
     public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
         registry.addHandler(notificationHandler, "/api/ws/notifications")
-                .setAllowedOrigins("*");  // Configure appropriately for production
-    }
-
-    // STOMP message broker for SimpMessagingTemplate
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry config) {
-        // Enable simple in-memory broker for topics
-        config.enableSimpleBroker("/topic", "/queue");
-        // Prefix for messages FROM client TO server
-        config.setApplicationDestinationPrefixes("/app");
-    }
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        // STOMP endpoint for SockJS fallback
-        registry.addEndpoint("/api/ws/stomp")
-                .setAllowedOrigins("*")
-                .withSockJS();
-        // Plain WebSocket STOMP endpoint
-        registry.addEndpoint("/api/ws/stomp")
-                .setAllowedOrigins("*");
+                .setAllowedOriginPatterns("*");
     }
 }
